Avoid creating an expensive NAT Gateway
This means we have to run the task on a container in a public subnet
which increases the chance that that someone could gain access to the
container. However, I believe there is still a firewall in place and in
any case the container is only alive for a few minutes every week, so I
think this is safe enough.

An alternative would've been to use a NAT EC2 instance rather than a
fully-fledged NAT gateway. However, this would be entirely unmanaged and
therefore potentially require extra monitoring/maintenance.

Note that the ScheduledFargateTaskOnPublicSubnet class is a bit of a
bodge. I hope to open a PR against the AWS CDK repo to make this kind of
setup achievable via configuration alone.