Bump Dockerfile to get newer version of rclone
We were seeing errors from the `rclone sync` command in
local-image/home/backup.sh which I strongly suspect relate to this bug
in rclone [1]. It looks as if this bug was fixed in this commit [2]
which was released in rclone v1.56.0.

The ECS task uses a Docker image which is implicitly built and pushed to
ECR as part of running the `cdk deploy` command. However, if nothing in
the Dockerfile or associated files has changed, the docker image will
not be re-built or pushed to ECR. Thus the same (old) version of rclone
will be used every time.

This commit splits the original RUN command in the Dockerfile into two
RUN commands as a simple way to force a re-build of the Docker image
thus installing a more recent version of rclone (v1.58.1).

This isn't ideal, because following this approach we'll have to come up
with ever more creative ways to keep upgrading rclone. However, we can
tackle that issue separately.

[1]: https://github.com/rclone/rclone/issues/5455
[2]: https://github.com/rclone/rclone/commit/94b1439299e57f84dec27e865fba2b7356f3ea79